requirements.txt
requests
pytest
python-dotenv
urllib3